@@ -5,105 +5,102 @@ use std::{env::Args, io};
 pub struct Config {
     /// The name of the program for usage when printing help information
     ///
-    pub program_name: String, 
+    pub program_name: String,
 
-    /// The name of the file assumed to contain the entry point 
+    /// The name of the file assumed to contain the entry point
     ///
-    pub file: String, 
+    pub file: String,
 
     /// The verbosity of the compiler. This is used to toggle debug information while the compiler
     /// is running
     ///
     pub verbose: bool,
 
-    /// This is a list of paths that are used to lookup imports 
+    /// This is a list of paths that are used to lookup imports
     ///
-    pub import_paths: Vec<String>, 
+    pub import_paths: Vec<String>,
 
     /// A copy of all the command line args to be parsed
     ///
-    pub args: Vec<String>, 
+    pub args: Vec<String>,
 
-    /// Run code after JIT compilation 
+    /// Run code after JIT compilation
     ///
     pub run: bool,
 
     /// Print the ir after compilation
     ///
-    pub ir: bool, 
+    pub ir: bool,
 
     /// Print module definitions
     ///
-    pub defs: bool, 
+    pub defs: bool,
 }
-
 
 impl Config {
     fn new() -> Self {
         Self {
-            program_name: String::new(), 
-            file: String::new(), 
-            verbose: false, 
-            import_paths: Vec::new(), 
-            args: Config::args(), 
+            program_name: String::new(),
+            file: String::new(),
+            verbose: false,
+            import_paths: Vec::new(),
+            args: Config::args(),
             run: false,
-            ir: false, 
-            defs: false, 
+            ir: false,
+            defs: false,
         }
     }
 
     fn set_verbose(&mut self, v: bool) {
-        self.verbose = v; 
+        self.verbose = v;
     }
 
     fn set_run(&mut self, v: bool) {
-        self.run = v; 
+        self.run = v;
     }
 
     fn set_ir(&mut self, v: bool) {
-        self.ir = v; 
+        self.ir = v;
     }
 
     fn set_defs(&mut self, v: bool) {
-        self.defs = v; 
+        self.defs = v;
     }
 
-
     fn init(&mut self) {
-
         let init_message = r#"
     Zulu programming language project setup 
-        "#; 
+        "#;
 
-        println!("{}", init_message); 
+        println!("{}", init_message);
         println!("Enter entry point (default: Main.zulu): ");
         let entry = io::read_to_string(io::stdin()).unwrap();
         println!("{:?}", entry);
     }
 
-    /// Parse command line arguments and create a config. 
+    /// Parse command line arguments and create a config.
     pub fn parse() -> Self {
-        let mut conf = Config::new(); 
-        let name = conf.clone().parse_name(); 
-        conf.program_name = name; 
-        let len = conf.args.len(); 
+        let mut conf = Config::new();
+        let name = conf.clone().parse_name();
+        conf.program_name = name;
+        let len = conf.args.len();
         if len == 1 {
             conf.clone().report("no input file provided".to_string());
         }
 
-        let mut c = conf.clone(); 
-        let args = conf.args.get(1 .. len).unwrap();
+        let mut c = conf.clone();
+        let args = conf.args.get(1..len).unwrap();
         let args = args.to_vec();
-        let len = args.len(); 
+        let len = args.len();
         let mut held = false;
-        let mut collect_paths = false; 
-        let mut collect_file = false; 
+        let mut collect_paths = false;
+        let mut collect_file = false;
         for (index, arg) in args.clone().into_iter().enumerate() {
             if arg.starts_with("--") {
                 if held {
-                    held = false; 
+                    held = false;
                     if collect_paths {
-                        collect_paths = false; 
+                        collect_paths = false;
                     }
                 }
                 match arg.as_str() {
@@ -124,37 +121,38 @@ impl Config {
                         c.set_verbose(true);
                     }
                     "--version" => {
-                        c.clone().report(format!("{} version 0.1.1", c.program_name.clone())); 
-                    } 
+                        c.clone()
+                            .report(format!("{} version 0.1.1", c.program_name.clone()));
+                    }
                     "--help" => {
-                        c.clone().report(c.clone().help().to_string()); 
-                    } 
-                     "--init" => {
-                        c.clone().init(); 
+                        c.clone().report(c.clone().help().to_string());
+                    }
+                    "--init" => {
+                        c.clone().init();
                     }
                     "--paths" => {
                         held = true;
-                        collect_paths = true; 
+                        collect_paths = true;
                     }
                     d => {
-                        c.clone().report(format!("invalid argument `{}` provided", d));
+                        c.clone()
+                            .report(format!("invalid argument `{}` provided", d));
                     }
                 }
             } else if arg.as_str() == "-" {
-                held = false; 
-                collect_paths = false; 
-                collect_file = true; 
-            }
-            else if index < len  && held {
+                held = false;
+                collect_paths = false;
+                collect_file = true;
+            } else if index < len && held {
                 if collect_paths {
-                    c = c.clone(); 
+                    c = c.clone();
                     c.import_paths.push(arg);
-                    c = c.clone(); 
+                    c = c.clone();
                 }
             } else if collect_file {
-                c.file = arg; 
+                c.file = arg;
             } else {
-                c .file = arg;
+                c.file = arg;
             }
         }
 
@@ -166,17 +164,16 @@ impl Config {
     }
 
     fn parse_name(self) -> String {
-        let name = self.args.first().unwrap().clone(); 
+        let name = self.args.first().unwrap().clone();
         let names_split: Vec<&str> = name.split("/").collect();
-        let names_split = names_split.to_vec(); 
+        let names_split = names_split.to_vec();
         names_split.last().unwrap().to_string()
     }
 
-    
-    
     fn help(self) -> String {
-        let name = self.parse_name(); 
-format!(r#"
+        let name = self.parse_name();
+        format!(
+            r#"
 usage: {} [options] - <file>
     |  {} <file>
     
@@ -189,12 +186,17 @@ usage: {} [options] - <file>
         --version               Shows the version of the program
         --help                  Prints this help file
 
-"#, name.clone(), name).trim_start().to_string()
+"#,
+            name.clone(),
+            name
+        )
+        .trim_start()
+        .to_string()
     }
 
     /// report a Config error
     pub fn report(self, message: String) {
-        println!("{}", message); 
-        std::process::exit(1); 
+        println!("{}", message);
+        std::process::exit(1);
     }
 }
