@@ -1,4 +1,4 @@
-use crate::ir::raw::RawFunction;
+use crate::ir::raw::TopLevel;
 use nom::branch::alt;
 use nom::bytes::complete::{tag, take_until, take_while, take_while1};
 use nom::combinator::all_consuming;
@@ -11,14 +11,14 @@ mod expression;
 mod function;
 mod types;
 
-use function::function;
+use self::function::top_levels;
 
 fn sp(i: &str) -> IResult<&str, &str, VerboseError<&str>> {
     let chars = " \t\r\n";
     take_while(move |c| chars.contains(c))(i)
 }
 
-fn identifier(i: &str) -> IResult<&str, &str, VerboseError<&str>> {
+pub fn identifier(i: &str) -> IResult<&str, &str, VerboseError<&str>> {
     take_while1(move |c: char| c.is_alphanumeric() || "_'".contains(c))(i)
 }
 
@@ -40,7 +40,7 @@ pub fn comments_ommited(i: &str) -> IResult<&str, String, VerboseError<&str>> {
     Ok((i, r.join("")))
 }
 
-pub fn program(i: &str) -> IResult<&str, Vec<RawFunction>, VerboseError<&str>> {
-    let (i, (_, funs, _)) = all_consuming(tuple((sp, separated_list0(sp, function), sp)))(&i)?;
+pub fn program(i: &str) -> IResult<&str, Vec<TopLevel>, VerboseError<&str>> {
+    let (i, (_, funs, _)) = all_consuming(tuple((sp, separated_list0(sp, top_levels), sp)))(&i)?;
     Ok((i, funs))
 }
