@@ -1,7 +1,8 @@
-use super::sp;
+use super::{identifier, sp};
 use crate::ir::Type;
 use nom::branch::alt;
 use nom::bytes::complete::{tag, take_while1};
+use nom::multi::separated_list0;
 use nom::error::VerboseError;
 use nom::sequence::tuple;
 use nom::IResult;
@@ -22,16 +23,8 @@ fn array_type(i: &str) -> IResult<&str, Type, VerboseError<&str>> {
 }
 
 fn list_type(i: &str) -> IResult<&str, Type, VerboseError<&str>> {
-
-    let (i, (_, _, _, _, t, _, _)) = tuple((
-        tag("List"), 
-        sp, 
-        tag("["), 
-        sp, 
-        type_literal, 
-        sp, 
-        tag("]")
-    ))(i)?;
+    let (i, (_, _, _, _, t, _, _)) =
+        tuple((tag("List"), sp, tag("["), sp, type_literal, sp, tag("]")))(i)?;
 
     Ok((i, Type::List(Box::new(t))))
 }
@@ -77,7 +70,14 @@ fn int128_type(i: &str) -> IResult<&str, Type, VerboseError<&str>> {
 }
 
 fn int_types(i: &str) -> IResult<&str, Type, VerboseError<&str>> {
-    alt((int32_type, int64_type, int8_type, int16_type, int128_type, int_type))(i)
+    alt((
+        int32_type,
+        int64_type,
+        int8_type,
+        int16_type,
+        int128_type,
+        int_type,
+    ))(i)
 }
 
 fn float_type(i: &str) -> IResult<&str, Type, VerboseError<&str>> {
@@ -104,6 +104,62 @@ fn str_type(i: &str) -> IResult<&str, Type, VerboseError<&str>> {
     let (i, _) = tag("String")(i)?;
     Ok((i, Type::String))
 }
+
+fn any_type(i: &str) -> IResult<&str, Type, VerboseError<&str>> {
+    let (i, _) = tag("Any")(i)?;
+    Ok((i, Type::Any))
+}
+
+fn user_type(i: &str) -> IResult<&str, Type, VerboseError<&str>> {
+    let (i, t) = tuple((identifier,))(i)?;
+    let (t,) = t;
+
+    Ok((i, Type::UserType(t.to_string())))
+}
+
+fn enum_type(i: &str) -> IResult<&str, Type, VerboseError<&str>> {
+    let (i, (parent, _, _, _, child)) = tuple((
+        identifier,
+        sp, 
+        tag("."), 
+        sp, 
+        identifier))(i)?;
+
+    Ok((i, Type::EnumType(parent.to_string(), child.to_string())))
+}
+
+pub fn lambda(i: &str) -> IResult<&str, Type, VerboseError<&str>> {
+    let (i, (_, _, _, args, _, _, _, _, ret)) = tuple((
+        tag("fn"),
+        sp,
+        tag("("),
+        separated_list0(tuple((sp, tag(","), sp)), type_literal),
+        tag(")"),
+        sp,
+        tag(":"),
+        sp,
+        type_literal
+    ))(i)?;
+
+    Ok((i, Type::Lambda(Box::new(ret), args.clone())))
+}
+
+
 pub fn type_literal(i: &str) -> IResult<&str, Type, VerboseError<&str>> {
-    alt((str_type, unit_type, usz_type, double_type, list_type, char_type, float_type, int_types, bool_type, array_type))(i)
+    alt((
+        lambda,
+        enum_type,
+        any_type,
+        str_type,
+        unit_type,
+        usz_type,
+        double_type,
+        list_type,
+        char_type,
+        float_type,
+        int_types,
+        bool_type,
+        array_type,
+        user_type,
+    ))(i)
 }
